+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Binary data escaping in `listener` module (enhancing `Get Binary File` keyword logging), by @HardNorth
+### Changed
+- Client version updated on [5.5.5](https://github.com/reportportal/client-Python/releases/tag/5.5.5), by @HardNorth
+
+## [5.5.1]
 ### Changed
 - Unified ReportPortal product spelling, by @HardNorth
 - Client version updated on [5.5.4](https://github.com/reportportal/client-Python/releases/tag/5.5.4), by @HardNorth